Create taxonomy-dropdown.php
Shortcode to create a dropdown for initial filtering of posts. [show_tax_dropdown tax="category" text="Choose..." excl=""]
tax is the name of the taxonomy (category, post_tag)
text is the first words available in the dropdown
excl is any terms (categories, tags) you wish to exclude. Provide the ids in a list seperated by spaces (33 2 4)